Rename main-window to main
