+++ src/main/listenForKeyboardShortcut.ts
