+++ src/main/saveBeforeClosing.ts
