+++ src/menu/shared-menu.ts
@@ -7,7 +7,7 @@ import {
   MenuItem,
 } from "electron"
 import { readFileSync } from "fs"
-import {} from "../main-window/"
+import {} from "../main/"
 
 import {
   userHasChosenMediaFile,
@@ -15,7 +15,7 @@ import {
   userWantsToSaveTranscript,
   closeTheWindow,
 } from "../app/ipcChannelNames"
-import { promptUserToSelectFile } from "../main-window/selectFile"
+import { promptUserToSelectFile } from "../main/selectFile"
 
 export function createSharedMenuItems(window: BrowserWindow) {
   const visit: MenuItemConstructorOptions = {
